Use a more efficient algorithm for list shuffling
Previously, the Yates-Fisher shuffling algorithm was used for all the
sequences, but unfortunatelly when applied to lists this algorithm is
O(N*N).

This patch implements a list shuffling algorithm that is O(N*log(N))
whitout using any extra heap memory.

It works by dividing the list to shuffle in two, randomly
redistributing the list elements between the two parts in-place and
recursing.

It is quite similar to quicksort with a random comparator but ensuring
than the list is always divided in two parts of (almost) equal size,
and so avoiding the O(N*N) worst case of quicksort.